Moving DefaultConfiguration to impl package
